+++ src/main/java/br/fatec/TemosVagas/entities/Empresa.java
@@ -1,7 +1,7 @@
 package br.fatec.TemosVagas.entities;
 
 import br.fatec.TemosVagas.entities.usuario.Usuario;
-import br.fatec.TemosVagas.entities.usuario.UsuarioRole;
+import br.fatec.TemosVagas.entities.enums.UsuarioRole;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.PrePersist;
